% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Politician.R
\docType{class}
\name{parliamentAPI-class}
\alias{parliamentAPI-class}
\alias{parliamentAPI}
\title{This class contain an API call to fetch Swedish Parliament members details.}
\arguments{
\item{startD}{startDate}

\item{endD}{endDate}
}
\value{
list of Politicians based on \code{startD} and \code{endD}
}
\description{
This class contain an API call to fetch Swedish Parliament members details.
}
\details{
Swedish Parliament API is called in order to fetch members details from 1990 to current date
Same Class is furthur used in making Shiny Dashborad to show details on UI
}

\examples{
\dontrun{
a = parliamentAPI()
a$CalData("2001-01-01" , "2010-01-01")
}
}
