\name{swedishParlimentAPI-package}
\alias{swedishParlimentAPI-package}
\alias{swedishParlimentAPI}
\docType{package}
\title{
\packageTitle{swedishParlimentAPI}
}
\description{
\packageDescription{swedishParlimentAPI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{swedishParlimentAPI}
\packageIndices{swedishParlimentAPI}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{swedishParlimentAPI}

Maintainer: \packageMaintainer{swedishParlimentAPI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
